+++ Packages/com.hung.base/Editor/Items/IItemIdEditorSource.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace Hung.Base.Editor
-{
-    public interface IItemIdEditorSource
-    {
-        IEnumerable<ItemIdEditorOption> GetOptions();
-    }
-}