+++ Packages/com.hung.base/Runtime/Gameplay/CharacterContracts/ILogicEffectReceiver.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace Hung.Base
-{
-    public interface ILogicEffectReceiver
-    {
-        float TakeDamage(EFFECT effect, float time, Vector3 sourcePos, float damage = 1, object source = null);
-    }
-}