+++ Packages/com.hung.base/Runtime/Gameplay/CharacterContracts/IProjectile.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace Hung.Base
-{
-    public interface IProjectile
-    {     
-        public void Projectile(Vector3 direction, float speed);
-    }
-}