+++ Packages/com.hung.base/Runtime/Services/Contracts/Input/Locator.Input.cs
@@ -1,7 +1,0 @@
-namespace Hung.Base
-{
-    public static partial class Locator
-    {
-        public static IInputService Input { get; set; }
-    }
-}