+++ Packages/com.hung.designpattern/Runtime/Observe/IObserver.cs
@@ -1,7 +1,0 @@
-﻿namespace DesignPattern
-{
-    public interface IObserver
-    {
-        public void OnNotify();
-    }
-}