+++ Packages/com.hung.ui/Runtime/Core/IUIPrefabProvider.cs
@@ -1,7 +1,0 @@
-namespace Hung.UI
-{
-    public interface IUIPrefabProvider
-    {
-        T GetPrefab<T>() where T : UICanvas;
-    }
-}